lisätty lukuun tasosysteemi.

+++ MathGame.Web/Models/MathWar/SaveData.cs
@@ -2,7 +2,7 @@ namespace MathGame.Web.Models.MathWar;
 
 public class SaveData
 {
-    public int HighScore { get; set; }
+    public long HighScore { get; set; }
     public string BestRating { get; set; } = "";
     public int BestLevel { get; set; }
 }
